Improvemetns in the code-
There is one issue which I have rectified:
a) In the `_createNewBattle` function in the Arena abstract contract, the line Battle storage` newBattle = _battles.push();` is missing an argument to push. The `push` function requires an argument of the struct type being stored in the array, which is Battle in this case. The correct line of code should be `Battle storage newBattle = _battles.push(Battle());` to create a new Battle struct and push it to the _battles array.